requirements.txt
py-cord
python-dotenv
requests
Pillow